+++ validator/db/iface/BUILD.bazel
@@ -6,9 +6,5 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/validator/db/iface",
     # Other packages must use github.com/prysmaticlabs/prysm/validator/db.Database alias.
     visibility = ["//validator/db:__subpackages__"],
-    deps = [
-        "//proto/slashing:go_default_library",
-        "//validator/db/kv:go_default_library",
-        "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
-    ],
+    deps = ["//validator/db/kv:go_default_library"],
 )
