+++ validator/db/iface/interface.go
@@ -5,8 +5,6 @@ import (
 	"context"
 	"io"
 
-	"github.com/prysmaticlabs/go-bitfield"
-	slashpb "github.com/prysmaticlabs/prysm/proto/slashing"
 	"github.com/prysmaticlabs/prysm/validator/db/kv"
 )
 
@@ -22,10 +20,6 @@ type ValidatorDB interface {
 	SaveGenesisValidatorsRoot(ctx context.Context, genValRoot []byte) error
 
 	// Proposer protection related methods.
-	ProposalHistoryForEpoch(ctx context.Context, publicKey []byte, epoch uint64) (bitfield.Bitlist, error)
-	SaveProposalHistoryForEpoch(ctx context.Context, publicKey []byte, epoch uint64, history bitfield.Bitlist) error
-
-	// New data structure methods
 	HighestSignedProposal(ctx context.Context, publicKey [48]byte) (uint64, error)
 	LowestSignedProposal(ctx context.Context, publicKey [48]byte) (uint64, error)
 	ProposalHistoryForSlot(ctx context.Context, publicKey [48]byte, slot uint64) ([32]byte, bool, error)
@@ -33,14 +27,10 @@ type ValidatorDB interface {
 	ProposedPublicKeys(ctx context.Context) ([][48]byte, error)
 
 	// Attester protection related methods.
-	AttestationHistoryForPubKeys(ctx context.Context, publicKeys [][48]byte) (map[[48]byte]*slashpb.AttestationHistory, error)
-	SaveAttestationHistoryForPubKeys(ctx context.Context, historyByPubKey map[[48]byte]*slashpb.AttestationHistory) error
 	HighestSignedTargetEpoch(ctx context.Context, publicKey [48]byte) (uint64, error)
 	HighestSignedSourceEpoch(ctx context.Context, publicKey [48]byte) (uint64, error)
 	SaveHighestSignedTargetEpoch(ctx context.Context, publicKey [48]byte, epoch uint64) error
 	SaveHighestSignedSourceEpoch(ctx context.Context, publicKey [48]byte, epoch uint64) error
-
-	// New attestation store methods.
 	AttestationHistoryForPubKeysV2(ctx context.Context, publicKeys [][48]byte) (map[[48]byte]kv.EncHistoryData, error)
 	SaveAttestationHistoryForPubKeysV2(ctx context.Context, historyByPubKeys map[[48]byte]kv.EncHistoryData) error
 	SaveAttestationHistoryForPubKeyV2(ctx context.Context, pubKey [48]byte, history kv.EncHistoryData) error
