+++ validator/db/kv/BUILD.bazel
@@ -4,12 +4,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "attestation_history.go",
         "attestation_history_v2.go",
         "db.go",
         "genesis.go",
-        "manage.go",
-        "proposal_history.go",
+        "historical_attestations.go",
         "proposal_history_v2.go",
         "schema.go",
     ],
@@ -17,15 +15,11 @@ go_library(
     visibility = ["//validator:__subpackages__"],
     deps = [
         "//beacon-chain/core/helpers:go_default_library",
-        "//proto/slashing:go_default_library",
         "//shared/bytesutil:go_default_library",
         "//shared/fileutil:go_default_library",
         "//shared/params:go_default_library",
-        "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
-        "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
-        "@com_github_wealdtech_go_bytesutil//:go_default_library",
         "@io_etcd_go_bbolt//:go_default_library",
         "@io_opencensus_go//trace:go_default_library",
     ],
@@ -34,24 +28,17 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "attestation_history_test.go",
         "attestation_history_v2_test.go",
         "db_test.go",
         "genesis_test.go",
-        "manage_test.go",
-        "proposal_history_test.go",
+        "historical_attestations_test.go",
         "proposal_history_v2_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
-        "//beacon-chain/core/helpers:go_default_library",
-        "//proto/slashing:go_default_library",
         "//shared/bytesutil:go_default_library",
         "//shared/params:go_default_library",
         "//shared/testutil/assert:go_default_library",
         "//shared/testutil/require:go_default_library",
-        "@com_github_pkg_errors//:go_default_library",
-        "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
-        "@io_etcd_go_bbolt//:go_default_library",
     ],
 )
