+++ validator/db/kv/db.go
@@ -109,21 +109,17 @@ func NewKVStore(dirPath string, pubKeys [][48]byte) (*Store, error) {
 	return kv, err
 }
 
-// GetKVStore returns the validator boltDB key-value store from directory. Returns nil if no such store exists.
-func GetKVStore(directory string) (*Store, error) {
-	fileName := filepath.Join(directory, ProtectionDbFileName)
-	if _, err := os.Stat(fileName); os.IsNotExist(err) {
-		return nil, nil
-	}
-	boltDb, err := bolt.Open(fileName, params.BeaconIoConfig().ReadWritePermissions, &bolt.Options{Timeout: params.BeaconIoConfig().BoltTimeout})
-	if err != nil {
-		if errors.Is(err, bolt.ErrTimeout) {
-			return nil, errors.New("cannot obtain database lock, database may be in use by another process")
+// UpdatePublicKeysBuckets for a specified list of keys.
+func (store *Store) UpdatePublicKeysBuckets(pubKeys [][48]byte) error {
+	return store.update(func(tx *bolt.Tx) error {
+		bucket := tx.Bucket(newHistoricProposalsBucket)
+		for _, pubKey := range pubKeys {
+			if _, err := bucket.CreateBucketIfNotExists(pubKey[:]); err != nil {
+				return errors.Wrap(err, "failed to create proposal history bucket")
+			}
 		}
-		return nil, err
-	}
-
-	return &Store{db: boltDb, databasePath: directory}, nil
+		return nil
+	})
 }
 
 // Size returns the db size in bytes.
