+++ validator/db/kv/proposal_history_v2_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"testing"
 
-	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
@@ -153,75 +152,6 @@ func TestPruneProposalHistoryBySlot_OK(t *testing.T) {
 			require.DeepEqual(t, signedRoot, sr[:], "Unexpected difference in bytes for epoch %d", slot)
 		}
 	}
-}
-
-func TestStore_ImportProposalHistory(t *testing.T) {
-	pubkey := [48]byte{3}
-	ctx := context.Background()
-	db := setupDB(t, [][48]byte{pubkey})
-	proposedSlots := make(map[uint64]bool)
-	proposedSlots[0] = true
-	proposedSlots[1] = true
-	proposedSlots[20] = true
-	proposedSlots[31] = true
-	proposedSlots[32] = true
-	proposedSlots[33] = true
-	proposedSlots[1023] = true
-	proposedSlots[1024] = true
-	proposedSlots[1025] = true
-	lastIndex := 1025 + params.BeaconConfig().SlotsPerEpoch
-
-	for slot := range proposedSlots {
-		slotBitlist, err := db.ProposalHistoryForEpoch(context.Background(), pubkey[:], helpers.SlotToEpoch(slot))
-		require.NoError(t, err)
-		slotBitlist.SetBitAt(slot%params.BeaconConfig().SlotsPerEpoch, true)
-		err = db.SaveProposalHistoryForEpoch(context.Background(), pubkey[:], helpers.SlotToEpoch(slot), slotBitlist)
-		require.NoError(t, err)
-	}
-	err := db.MigrateV2ProposalFormat(ctx)
-	require.NoError(t, err)
-
-	for slot := uint64(0); slot <= lastIndex; slot++ {
-		if _, ok := proposedSlots[slot]; ok {
-			root, _, err := db.ProposalHistoryForSlot(ctx, pubkey, slot)
-			require.NoError(t, err)
-			require.DeepEqual(t, bytesutil.PadTo([]byte{1}, 32), root[:], "slot: %d", slot)
-			continue
-		}
-		root, _, err := db.ProposalHistoryForSlot(ctx, pubkey, slot)
-		require.NoError(t, err)
-		require.DeepEqual(t, bytesutil.PadTo([]byte{}, 32), root[:])
-	}
-}
-
-func TestShouldImportProposals(t *testing.T) {
-	pubkey := [48]byte{3}
-	db := setupDB(t, nil)
-
-	shouldImport, err := db.shouldImportProposals()
-	require.NoError(t, err)
-	require.Equal(t, false, shouldImport, "Empty bucket should not be imported")
-	err = db.OldUpdatePublicKeysBuckets([][48]byte{pubkey})
-	require.NoError(t, err)
-	shouldImport, err = db.shouldImportProposals()
-	require.NoError(t, err)
-	require.Equal(t, true, shouldImport, "Bucket with content should be imported")
-}
-
-func TestStore_UpdateProposalsProtectionDb(t *testing.T) {
-	pubkey := [48]byte{3}
-	db := setupDB(t, [][48]byte{pubkey})
-	ctx := context.Background()
-	err := db.OldUpdatePublicKeysBuckets([][48]byte{pubkey})
-	require.NoError(t, err)
-	shouldImport, err := db.shouldImportProposals()
-	require.NoError(t, err)
-	require.Equal(t, true, shouldImport, "Bucket with content should be imported")
-	err = db.MigrateV2ProposalsProtectionDb(ctx)
-	require.NoError(t, err)
-	shouldImport, err = db.shouldImportProposals()
-	require.NoError(t, err)
-	require.Equal(t, false, shouldImport, "Proposals should not be re-imported")
 }
 
 func TestStore_ProposedPublicKeys(t *testing.T) {
